+++ src/Components/Validator/Model/ValidationResult.php
@@ -6,15 +6,24 @@ class ValidationResult
 {
 
     /**
+     * @var ValidationTest[]
+     */
+    private $tests;
+
+
+    /**
      * @var ValidationError[]
      */
     private $errors;
 
+    
     /**
+     * @param ValidationTest[] $tests
      * @param ValidationError[] $errors
      */
-    public function __construct(array $errors)
+    public function __construct(array $tests, array $errors)
     {
+        $this->tests = $tests;
         $this->errors = $errors;
     }
 
@@ -24,6 +33,14 @@ class ValidationResult
     public function isValid(): bool
     {
         return count($this->errors) === 0;
+    }
+
+    /**
+     * @return ValidationTest[]
+     */
+    public function getTests(): array
+    {
+        return $this->tests;
     }
 
     /**